refactor all remaining things
