+++ tests/utils.py
@@ -1,6 +1,6 @@
 import ast
 
-from personal_python_ast_optimizer.parser import run_minify_parser
+from personal_python_ast_optimizer.parser.run import run_minify_parser
 from personal_python_ast_optimizer.parser.config import (
     SectionsToSkipConfig,
     SkipConfig,
